simple-hb written but untested

+++ 3nodes/LANChat-Containers/init.sh
@@ -2,9 +2,15 @@
 ### Instantiate 3 docker containers wth LANChat running
 ###
 
+cleanup() {
+  docker stop peer1 peer2 peer3
+  docker rm peer1 peer2 peer3
+  docker network rm lanchat-br 2> /dev/null
+}
 
-
-cat << EOF > Dockerfile
+if [ ! -f "./Dockerfile" ]
+then
+  cat << EOF > Dockerfile
 # Pull base image.
 FROM python:3
 
@@ -17,6 +23,9 @@ CMD ./main.py
 # Expose ports.
 EXPOSE 8080
 EOF
+fi
+
+echo Dockerfile written
 
 if [ ! -d "./LANChat" ]
 then
@@ -40,13 +49,6 @@ EOF
   echo host configuration written
 fi
 
-
-cleanup() {
-  docker stop peer1 peer2 peer3
-  docker rm peer1 peer2 peer3
-  docker network rm lanchat-br 2> /dev/null
-}
-
 cleanup
 
 docker build -t lanchat .
@@ -55,23 +57,3 @@ docker network create \
   -d bridge \
   --subnet=10.0.3.0/24 \
   lanchat-br
-
-docker run -it --name peer1 \
-  --network lanchat-br \
-  --ip "10.0.3.21" \
-  --publish 8080:8080 \
-  lanchat
-
-# # Run this in another terminal
-docker run -it --name peer2 \
-  --network lanchat-br \
-  --ip "10.0.3.22" \
-  --publish 8081:8080 \
-  lanchat
-#
-# # Run this in another terminal
-# docker run -it --name peer3 \
-#   --network lanchat-br \
-#   --ip "10.0.3.23" \
-#   --publish 8082:8080 \
-#   lanchat
